Fix Flags Glitching
Trying to have our cake and eating it. Flags use bSimFall to smooth out
their movement, which causes Physics to be replicated, which makes them
assume Bases like lifts on landing from PHYS_Falling.
This commit adds a watchdog that resets the Bases of all flags every
tick if the flags Physics is either PHYS_None or PHYS_Falling.